+++ cubes/workarena/src/workarena_cube/__init__.py
@@ -1,6 +1,7 @@
 from workarena_cube.benchmark import WorkArenaBenchmark
 from workarena_cube.debug import CheatAgent, make_debug_agent, get_debug_benchmark
 from workarena_cube.task import WorkArenaTask, WorkArenaTaskConfig
+from workarena_cube.tools import WorkArenaBrowserTool, WorkArenaCheatTool, WorkArenaInfeasibleTool, WorkarenaBrowserToolConfig, WorkArenaInfeasibleToolConfig, WorkArenaCheatToolConfig
 
 __all__ = [
     "WorkArenaBenchmark",
@@ -9,4 +10,11 @@ __all__ = [
     "CheatAgent",
     "make_debug_agent",
     "get_debug_benchmark",
+    "WorkArenaBrowserTool",
+    "WorkArenaCheatTool",
+    "WorkArenaInfeasibleTool",
+    "WorkarenaBrowserToolConfig",
+    "WorkArenaInfeasibleToolConfig",
+    "WorkArenaCheatToolConfig",
+
 ]
