+++ cubes/workarena/src/workarena_cube/debug.py
@@ -23,7 +23,6 @@ from cube.core import Action, ActionSchema, Observation
 from cube.testing import run_debug_suite
 
 from workarena_cube.benchmark import WorkArenaBenchmark
-from workarena_cube.task import WorkArenaTaskConfig
 from workarena_cube.tools import WorkArenaCheatToolConfig
 
 logger = logging.getLogger(__name__)
